# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Fut Match API",
        "description": "API para o sistema Fut Match - Plataforma de gerenciamento de partidas de futebol",
        "contact": {
            "email": "contato@futmatch.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor de desenvolvimento"
        },
        {
            "url": "https://back-end-production-c28b.up.railway.app",
            "description": "Servidor de produção"
        }
    ],
    "paths": {
...
        "/email/verify/{id}/{hash}": {
            "get": {
                "tags": [
                    "Email Verification"
                ],
                "summary": "Verifica email através do link",
                "description": "Verificar email através do link enviado",
                "operationId": "80503b7246cd847e6f41c7dcd2797a98",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "hash",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Email verificado com sucesso"
                    },
                    "400": {
                        "description": "Link inválido ou expirado"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum Token Authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}